requirements.txt
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
filelock==3.16.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
jiter==0.5.0
JPype1==1.5.0
konlpy==0.6.0
lxml==5.3.0
numpy==2.1.1
packaging==24.1
portalocker==2.10.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-mecab-ko==1.3.7
python-mecab-ko-dic==2.1.1.post2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
sacrebleu==2.4.3
sniffio==1.3.1
tabulate==0.9.0
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
